🔧 refactor: restructure authentication and search modules
crawjud/addons/auth/__init__.py: Removed unused code and added ProjudiAuth to the module exports.
crawjud/addons/auth/controller.py: Created a new controller for authentication operations.
crawjud/addons/auth/elaw.py: Updated ElawAuth class to remove unnecessary name parameter.
crawjud/addons/auth/esaj.py: Updated EsajAuth class to remove unnecessary name parameter.
crawjud/addons/auth/pje.py: Updated PjeAuth class to remove unnecessary name parameter.
crawjud/addons/auth/projudi.py: Added ProjudiAuth class for authentication.
crawjud/addons/search/__init__.py: Removed unused code and added ProjudiSearch to the module exports.
crawjud/addons/search/controller.py: Created a new controller for search operations.
crawjud/addons/search/elaw.py: Updated ElawSearch class to remove unnecessary name parameter.
crawjud/addons/search/esaj.py: Updated EsajSearch class to remove unnecessary name parameter.
crawjud/addons/search/projudi.py: Added ProjudiSearch class for search functionality.
crawjud/core/master.py: Refactored row property to use a private variable for better encapsulation.
